fix(cleanup): general cleanup to remove not implemented things and also have better naming
