+++ metrics/AGENTS.md
@@ -66,7 +66,7 @@ All adapters invoke the shared bridge `python -m harness.hooks.bridge <event>
 ## Learned User Preferences
 
 - Write git commits using Conventional Commits (`type(scope): subject`, with optional body and footer).
-- When the user requests a staged-only commit (for example from a diff-tab flow), treat their staged file list as authoritative: commit only what is already staged and do not stage additional files.
+- When the user requests a staged-only commit (for example from a diff-tab flow), treat their staged file list as authoritative: commit only what is already staged and do not stage additional files; when handling subagent completion notifications, do not restate user-visible subagent output unless the user asks or cross-agent synthesis is required.
 - Before local cluster-backed or kubectl-driven checks, confirm the intended
   Kubernetes context is selected (for example
   `kubectl config use-context kind-metrics`) and that kind, Helm, and kubectl
