+++ metrics/src/metrics/schemas/__init__.py
@@ -1,35 +1,17 @@
 """Pydantic request, response, and internal transfer models."""
 
 from metrics.schemas.metrics import (
-    CapacityReading,
     ErrorDetail,
     ErrorResponse,
     PlatformMetricsData,
     PlatformMetricsResponse,
-    ResourceSnapshot,
     ResponseMetadata,
-    SessionMetricsData,
-    SessionMetricsResponse,
-    UsageReading,
-    UsageSnapshot,
-    UserMetricsData,
-    UserMetricsResponse,
-    UtilizationSnapshot,
 )
 
 __all__ = [
-    "CapacityReading",
     "ErrorDetail",
     "ErrorResponse",
     "PlatformMetricsData",
     "PlatformMetricsResponse",
-    "ResourceSnapshot",
     "ResponseMetadata",
-    "SessionMetricsData",
-    "SessionMetricsResponse",
-    "UsageReading",
-    "UsageSnapshot",
-    "UserMetricsData",
-    "UserMetricsResponse",
-    "UtilizationSnapshot",
 ]
