+++ metrics/src/metrics/schemas/metrics.py
@@ -5,7 +5,7 @@ from __future__ import annotations
 from datetime import UTC, datetime
 from typing import Literal
 
-from pydantic import BaseModel, ConfigDict, Field
+from pydantic import BaseModel, Field
 
 
 class ResponseMetadata(BaseModel):
@@ -32,70 +32,6 @@ class ErrorDetail(BaseModel):
     )
 
 
-class ResourceSnapshot(BaseModel):
-    """Resource quantity values represented in Kubernetes-friendly strings."""
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    cpu: str = Field(description="CPU quantity as a core string.", examples=["512"])
-    memory: str = Field(
-        description="Memory quantity string with GiB unit.",
-        examples=["2048 GiB"],
-    )
-    ephemeral_memory: str = Field(
-        default="0 GiB",
-        alias="ephemeral-memory",
-        description="Ephemeral memory quantity string with GiB unit.",
-        examples=["0 GiB"],
-    )
-    gpu: str = Field(
-        default="0",
-        description="GPU resource count as a string quantity.",
-        examples=["4"],
-    )
-
-
-class UtilizationSnapshot(BaseModel):
-    """Utilization ratios (0-1) for requested over capacity resources."""
-
-    model_config = ConfigDict(populate_by_name=True)
-
-    cpu: float = Field(
-        ge=0,
-        le=1,
-        description="CPU utilization ratio from 0 to 1.",
-        examples=[0.25],
-    )
-    memory: float = Field(
-        ge=0,
-        le=1,
-        description="Memory utilization ratio from 0 to 1.",
-        examples=[0.375],
-    )
-    ephemeral_memory: float = Field(
-        default=0.0,
-        alias="ephemeral-memory",
-        ge=0,
-        le=1,
-        description="Ephemeral memory utilization ratio from 0 to 1.",
-        examples=[0.0],
-    )
-    gpu: float = Field(
-        default=0.0,
-        ge=0,
-        le=1,
-        description="GPU utilization ratio from 0 to 1.",
-        examples=[0.0],
-    )
-
-
-class UsageSnapshot(BaseModel):
-    """Requested and utilization usage sections."""
-
-    requested: ResourceSnapshot
-    utilization: UtilizationSnapshot
-
-
 class PlatformMetricsData(BaseModel):
     """Platform metrics payload."""
 
@@ -120,27 +56,6 @@ class PlatformMetricsData(BaseModel):
     )
 
 
-class UserMetricsData(BaseModel):
-    """User metrics payload."""
-
-    scope: Literal["user"] = "user"
-    cluster: str
-    user_id: str
-    capacity: ResourceSnapshot
-    usage: UsageSnapshot
-
-
-class SessionMetricsData(BaseModel):
-    """Session metrics payload."""
-
-    scope: Literal["session"] = "session"
-    cluster: str
-    user_id: str
-    session_id: str
-    capacity: ResourceSnapshot
-    usage: UsageSnapshot
-
-
 class PlatformMetricsResponse(BaseModel):
     """Success envelope for platform metrics."""
 
@@ -152,32 +67,6 @@ class PlatformMetricsResponse(BaseModel):
     metadata: ResponseMetadata
     status: Literal["Success"] = "Success"
     data: PlatformMetricsData
-
-
-class UserMetricsResponse(BaseModel):
-    """Envelope for user metrics responses."""
-
-    version: str = Field(
-        description="Versioned API group identifier.",
-        examples=["metrics.canfar.net/v1"],
-    )
-    kind: Literal["UserMetrics"] = "UserMetrics"
-    metadata: ResponseMetadata
-    status: Literal["Success"] = "Success"
-    data: UserMetricsData
-
-
-class SessionMetricsResponse(BaseModel):
-    """Envelope for session metrics responses."""
-
-    version: str = Field(
-        description="Versioned API group identifier.",
-        examples=["metrics.canfar.net/v1"],
-    )
-    kind: Literal["SessionMetrics"] = "SessionMetrics"
-    metadata: ResponseMetadata
-    status: Literal["Success"] = "Success"
-    data: SessionMetricsData
 
 
 class ErrorResponse(BaseModel):
@@ -193,19 +82,3 @@ class ErrorResponse(BaseModel):
     error: ErrorDetail
 
 
-class CapacityReading(BaseModel):
-    """Internal model returned by capacity providers."""
-
-    cpu_cores: float
-    memory_gib: float
-    source: str
-    observed_at: datetime
-
-
-class UsageReading(BaseModel):
-    """Internal model returned by utilization providers."""
-
-    requested_cpu_cores: float
-    requested_memory_gib: float
-    source: str
-    observed_at: datetime
